// src/types/settings.ts
export interface ISettings {
    _id?: string;
    obligadoContabilidad: boolean;
    activo: boolean;
    razonSocial: string;
    nombreComercial: string;
    direccionMatriz: string;
    direccionEstablecimiento: string;
    contribuyenteRimpe: boolean;
    ruc: string;
    regimenMicroempresa: boolean;
}